Reverting r12315 (More Windows 64b compilation warning fixes) due to compilation failures
git-svn-id: http://skia.googlecode.com/svn/trunk@12316 2bbb7eff-a529-9590-31e7-b0007b416f81

+++ src/images/bmpdecoderhelper.h
@@ -72,7 +72,7 @@ class BmpDecoderHelper {
   BmpDecoderHelper() { }
   ~BmpDecoderHelper() { }
   bool DecodeImage(const char* data,
-                   size_t len,
+                   int len,
                    int max_pixels,
                    BmpDecoderCallback* callback);
 
@@ -91,7 +91,7 @@ class BmpDecoderHelper {
 
   const uint8* data_;
   int pos_;
-  size_t len_;
+  int len_;
   int width_;
   int height_;
   int bpp_;
